+++ core-ng/src/main/java/core/framework/api/http/HTTPClientBuilder.java
@@ -59,7 +59,7 @@ public final class HTTPClientBuilder {
             if (!enableCookie) builder.disableCookieManagement();
 
             CloseableHttpClient httpClient = builder.build();
-            return new HTTPClient(httpClient, slowOperationThreshold.toMillis());
+            return new HTTPClient(httpClient, slowOperationThreshold);
         } catch (NoSuchAlgorithmException | KeyManagementException | KeyStoreException e) {
             throw new Error(e);
         } finally {