+++ core-ng/src/main/java/core/framework/impl/queue/RabbitMQConsumer.java
@@ -32,16 +32,16 @@ public class RabbitMQConsumer implements Consumer, AutoCloseable {
     private final Queue<QueueingConsumer.Delivery> deliveries = new ConcurrentLinkedQueue<>();
     private final Channel channel;
     private final String queue;
-    private final long slowOperationThresholdInMs;
+    private final long slowOperationThresholdInNanos;
     private final Thread consumerThread;
     private volatile ShutdownSignalException shutdown;
     private volatile ConsumerCancelledException cancelled;
 
     // refer to com.rabbitmq.client.QueueingConsumer
-    public RabbitMQConsumer(Channel channel, String queue, int prefetchCount, long slowOperationThresholdInMs) {
+    public RabbitMQConsumer(Channel channel, String queue, int prefetchCount, long slowOperationThresholdInNanos) {
         this.channel = channel;
         this.queue = queue;
-        this.slowOperationThresholdInMs = slowOperationThresholdInMs;
+        this.slowOperationThresholdInNanos = slowOperationThresholdInNanos;
         try {
             channel.basicQos(prefetchCount);
             channel.basicConsume(queue, false, this);   // QOS only works with manual ack
@@ -149,7 +149,7 @@ public class RabbitMQConsumer implements Consumer, AutoCloseable {
             long elapsedTime = watch.elapsedTime();
             ActionLogContext.track("rabbitMQ", elapsedTime);
             logger.debug("acknowledge, queue={}, deliveryTag={}, multiple={}, elapsedTime={}", queue, deliveryTag, multiple, elapsedTime);
-            if (elapsedTime > slowOperationThresholdInMs) {
+            if (elapsedTime > slowOperationThresholdInNanos) {
                 logger.warn(Markers.errorCode("SLOW_RABBITMQ"), "slow rabbitMQ operation, elapsedTime={}", elapsedTime);
             }
         }