+++ log-processor/src/main/java/core/log/domain/ActionDocument.java
@@ -35,6 +35,8 @@ public class ActionDocument {
     public String errorMessage;
     @XmlElement(name = "elapsed")
     public Long elapsed;
+    @XmlElement(name = "cpu_time")
+    public Long cpuTime;
     @XmlElement(name = "context")
     public Map<String, String> context;
     @XmlElement(name = "perf_stats")